issue-634: refactor(savings): decompose savings_goals_screen into sub-files (#634)
Split the 1211-line savings_goals_screen.dart into focused files, all <500
lines, with no behavioural change (pure code move + privacy adjustments):

- lib/screens/savings_goal_detail_screen.dart  (SavingsGoalDetailScreen, 428)
- lib/widgets/savings/goal_card.dart            (GoalCard, 219)
- lib/widgets/savings/add_contribution_sheet.dart (AddContributionSheet, 168)
- lib/widgets/savings/how_it_works_card.dart    (HowItWorksCard, 97)
- lib/utils/savings_goal_format.dart            (shared deadline helpers, 20)
- savings_goals_screen.dart slimmed 1211 -> 314

Private classes referenced across the new file boundary were promoted to
public (Dart privacy is per-library); shared _deadlineLabel/_isOverdue
helpers extracted to a util so both the card and detail screen reuse them.

Part of #634 (savings goals screen; 4 large screens remain).
